compiler warning fix; strict scaling will respect FILTER and SMOOTH settings for resize
git-svn-id: https://svn.code.sf.net/p/cgiscaler/code/CGIScaler@21 61d5e512-982d-0410-aaff-b96e2116ef47

+++ src/serve.c
@@ -120,7 +120,7 @@ void serve_from_blob(unsigned char *blob, size_t blob_len, char *mime_type) {
 	debug(DEB,"Serving from BLOB: size: %d", blob_len);
 
 	printf("Content-Type: %s\n", mime_type);
-	printf("Content-Length: %u\n", blob_len);
+	printf("Content-Length: %u\n", (unsigned int) blob_len);
 
 	printf("\n");
 	/* flushing buffers befor we do direct fd write */
